meshio: move Mesh and ElementType to root module
Also

- don't implement Display directly on them, so that multiple formats can
  be displayed,
- update the naming in the FFI.
