+++ tools/src/bin/medit2svg.rs
@@ -2,8 +2,8 @@ use anyhow::Context as _;
 use anyhow::Result;
 use coupe::sprs::CsMatView;
 use coupe::Point2D;
-use mesh_io::medit::ElementType;
-use mesh_io::medit::Mesh;
+use mesh_io::ElementType;
+use mesh_io::Mesh;
 use rayon::iter::IntoParallelRefIterator as _;
 use rayon::iter::ParallelIterator as _;
 use std::collections::HashSet;
