+++ tools/src/bin/mesh-part.rs
@@ -7,8 +7,8 @@ use coupe::nalgebra::DefaultAllocator;
 use coupe::nalgebra::DimDiff;
 use coupe::nalgebra::DimSub;
 use coupe::nalgebra::ToTypenum;
-use mesh_io::medit::Mesh;
 use mesh_io::weight;
+use mesh_io::Mesh;
 use std::env;
 use std::fs;
 use std::io;
