+++ tools/src/bin/part-bench.rs
@@ -8,8 +8,8 @@ use coupe::nalgebra::DimDiff;
 use coupe::nalgebra::DimSub;
 use coupe::nalgebra::ToTypenum;
 use criterion::Criterion;
-use mesh_io::medit::Mesh;
 use mesh_io::weight;
+use mesh_io::Mesh;
 use std::env;
 use std::fs;
 use std::io;
